+++ Vault-Coder/Changes.md
@@ -4,9 +4,14 @@
 ### Breaking changes
 - Generated **UniqueXAccess** & **UniqueXAccessLike** traits now extend **UniqueModelAccess**, 
   which means that the abstract property is now `condition: Condition` instead of `accessCondition: Option[Condition]`
+### New features
+- Added support for targeting access -writing
+  - This feature is enabled by passing `-T` as a console argument
 ### Bugfixes
 - Fixed a bug that would cause an infinite loop when two classes directly reference each other
 - Fixed a bug that would cause certain inheritance-based combo classes to interrupt the code-generation process
+### Other changes
+- `schema` and `.apply(ModelLike)` in the data classes are no longer low merge priorities
 
 ## v1.12.1 - 23.1.2025
 A small update fixing some bugs